fix exro_explorer_transaction_link, re-arrange code;
