+++ core/src/cli_commands/update_pool.rs
@@ -26,8 +26,8 @@ use crate::{
         make_pool_box_candidate_unchecked, BallotBox, CastBallotBoxVoteParameters, PoolBox,
         PoolBoxWrapper, VoteBallotBoxWrapper,
     },
-    cli_commands::ergo_explorer_transaction_link,
     contracts::pool::PoolContract,
+    explorer_api::ergo_explorer_transaction_link,
     node_interface::{SignTransaction, SubmitTransaction},
     oracle_config::BASE_FEE,
     oracle_state::{OraclePool, PoolBoxSource, StageError, UpdateBoxSource, VoteBallotBoxesSource},
